Switch SizeHint fields to u64.

+++ tokio-buf/src/buf_stream/size_hint.rs
@@ -7,9 +7,9 @@
 /// * `None` for `upper`.
 #[derive(Debug, Default, Clone)]
 pub struct SizeHint {
-    available: usize,
-    lower: usize,
-    upper: Option<usize>,
+    available: u64,
+    lower: u64,
+    upper: Option<u64>,
 }
 
 impl SizeHint {
@@ -22,12 +22,12 @@ impl SizeHint {
     /// the `BufStream` without `NotReady` being returned.
     ///
     /// It is possible that more data is currently available.
-    pub fn available(&self) -> usize {
+    pub fn available(&self) -> u64 {
         self.available
     }
 
     /// Set the value of the `available` hint.
-    pub fn set_available(&mut self, value: usize) {
+    pub fn set_available(&mut self, value: u64) {
         self.available = value;
 
         if self.lower < value {
@@ -44,7 +44,7 @@ impl SizeHint {
 
     /// Returns the lower bound of data that the `BufStream` will yield before
     /// completing.
-    pub fn lower(&self) -> usize {
+    pub fn lower(&self) -> u64 {
         self.lower
     }
 
@@ -53,14 +53,14 @@ impl SizeHint {
     /// # Panics
     ///
     /// This function panics if `value` is less than `available`.
-    pub fn set_lower(&mut self, value: usize) {
+    pub fn set_lower(&mut self, value: u64) {
         assert!(value >= self.available);
         self.lower = value;
     }
 
     /// Returns the upper bound of data the `BufStream` will yield before
     /// completing, or `None` if the value is unknown.
-    pub fn upper(&self) -> Option<usize> {
+    pub fn upper(&self) -> Option<u64> {
         self.upper
     }
 
@@ -69,7 +69,7 @@ impl SizeHint {
     /// # Panics
     ///
     /// This function panics if `value` is less than `lower`.
-    pub fn set_upper(&mut self, value: usize) {
+    pub fn set_upper(&mut self, value: u64) {
         // There is no need to check `available` as that is guaranteed to be
         // less than or equal to `lower`.
         assert!(value >= self.lower, "`value` is less than than `lower`");
